projet_ia/requirements.txt
flask
flask-cors
torch
torchvision
pandas
joblib
requests
Pillow
scikit-learn
gunicorn
